+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for insanic
 ------------------
 
 - FIX: removes aws_xray_sdk from insanic testing helpers
+- CHORE: removes grpc install extras
 - UPDATE: remove service semaphore
 - UPDATE: changes pool size per host from 30 to 0