+++ CHANGELOG.md
@@ -5,7 +5,10 @@ Changelog for insanic
 0.8.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- FIX: removes aws_xray_sdk from insanic testing helpers
+- CHORE: removes grpc install extras
+- UPDATE: remove service semaphore
+- UPDATE: changes pool size per host from 30 to 0
 
 
 0.8.1 (2019-08-27)